Replace handler functions with parseMessage
This feels clunky for 2 reasons: 1) I'm making different names for the
same thing, to get around the error severity being listed in the middle
of the data I'm trying to parse, 2) though I have a _ case for the first
character, to handle lines that don't begin with I, W, or E, I don't
have anything beyond that to ensure that things parse properly; I'm just
hoping that the fields after those letters are correct. The Unknown
message type would seem to want to handle poorly formatted lines that
start with those letters, too, but right now that won't ever happen.