Refactor parseMessage in LogAnalysis.hs
This is an attempt to be simpler, more clear, and more concise. It's
still not at all robust, and in fact fails on line 4113 of the sample
error.log file, which just has "I 4993" on it (and thus nothing beyond
column 2, which my code can't handle). This does not feel like the way
to do this kind of stuff, which I think is setting me up for learning
about Haskell's parsers, which I've heard mention of on freenode, and
which I've heard Simon Peyton Jones claim as one of Haskell's strengths.