Add solution to Exercise 5
This one is a bit ambiguous. The assignment text talks about a stack
machine that works with addition, multiplication, and, and or, but then
it asks the student to make a compile function that does mathematical
compilation, which could include and and or, but then I'm supposed to
make this work with Expr, by instancing it for Program, but that would
mean I needed to add and and or to to my instance, which messes up all
the old instances. I'm just leaving it this way, and you can only work
with the stack with push, add, and mul. I get the idea.